requirements.txt
# Dependências principais
numpy>=1.21.0
scikit-learn>=1.0.0
regex>=2022.1.18
joblib>=1.1.0
tqdm>=4.62.0
pyyaml>=6.0

# Para desenvolvimento e testes
pytest>=7.0.0
black>=22.1.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.931

# Para notebooks
jupyter>=1.0.0
matplotlib>=3.5.0
pandas>=1.4.0 